requirements.txt
groq
requests
python-dotenv
Pillow
librosa
numpy
gradio